agentcore-gateway/lambda_to_mcp/requirements.txt
strands-agents
strands-agents-tools
uv
boto3
bedrock-agentcore
bedrock-agentcore-starter-toolkit
aws-cdk-lib
constructs